extracts the eventDetails into a seperate component
